Add gcc phat

+++ neural_tdoa/train.py
@@ -3,6 +3,7 @@ import shutil
 import torch
 
 from datasets.dataset import TdoaDataset
+from datasets.math_utils import compute_distance, gcc_phat, normalize_tdoa
 from neural_tdoa.metrics import Loss, average_rms_error
 from neural_tdoa.model import TdoaCrnn10
 
@@ -22,10 +23,10 @@ class LitTdoaCrnn10(pl.LightningModule):
 
     def training_step(self, batch, batch_idx):
         x, y = batch
-        x = self.model(x)
+        predictions = self.model(x)
 
-        loss = self.loss(x, y)
-        rms = average_rms_error(y, x)
+        loss = self.loss(predictions, y)
+        rms = average_rms_error(y, predictions)
 
         self.log("train_loss", loss)
         self.log("train_rms", rms, prog_bar=True)
@@ -33,15 +34,31 @@ class LitTdoaCrnn10(pl.LightningModule):
         return loss
     
     def validation_step(self, batch, batch_idx):
-        x, y = batch
-        x = self.model(x)
+        X, Y = batch
+        predictions = self.model(X)
 
-        loss = self.loss(x, y)
+        loss = self.loss(predictions, Y)
 
-        rms = average_rms_error(y, x)
+        rms = average_rms_error(predictions, X)
         self.log("validation_loss", loss, prog_bar=True)
         self.log("validation_rms", rms, prog_bar=True)
+
+        validation_config = self.config["validation_dataset"]
+        fs = validation_config["base_sampling_rate"]
         
+        tdoas_gcc_phat = torch.zeros_like(Y)
+        for i, x in enumerate(X):
+            tdoas_gcc_phat[i] = _compute_tdoa_with_gcc_phat(
+                                x[0],
+                                x[1],
+                                fs,
+                                validation_config["mic_coordinates"]
+                               )
+           
+        rms_gcc = average_rms_error(Y, tdoas_gcc_phat)
+            
+        self.log("validation_rms_gcc", rms_gcc, prog_bar=True)
+
     def configure_optimizers(self):
         lr = self.config["training"]["learning_rate"]
         optimizer = self.config["training"]["optimizer"]
@@ -58,7 +75,8 @@ def train_pl(config):
     dataset_train = TdoaDataset(config["training_dataset"])
     dataset_val = TdoaDataset(config["validation_dataset"])
 
-    batch_size = config["training"]["batch_size"]
+    training_config = config["training"]
+    batch_size = training_config["batch_size"]
     dataset_train = torch.utils.data.DataLoader(dataset_train,
                                                 batch_size=batch_size,
                                                 shuffle=True,
@@ -72,14 +90,24 @@ def train_pl(config):
                                               drop_last=False,
                                               num_workers=2)
 
-    num_epochs = config["training"]["num_epochs"]
     
     gpus = 1 if torch.cuda.is_available() else 0
     tb_logger = pl_loggers.TensorBoardLogger("logs/")
 
-    trainer = pl.Trainer(max_epochs=num_epochs, log_every_n_steps=10,
+    trainer = pl.Trainer(max_epochs=training_config["num_epochs"],
+                         log_every_n_steps=training_config["log_every_n_steps"],
                          gpus=gpus, logger=tb_logger)
     trainer.fit(model, dataset_train, val_dataloaders=dataset_val)
 
-    shutil.rmtree(config["training_dataset"]["dataset_dir"])
-    shutil.rmtree(config["validation_dataset"]["dataset_dir"])
+    if training_config["delete_datasets_after_training"]:
+        shutil.rmtree(config["training_dataset"]["dataset_dir"])
+        shutil.rmtree(config["validation_dataset"]["dataset_dir"])
+
+
+def _compute_tdoa_with_gcc_phat(x1, x2, fs, mic_positions):
+    mic_distances = compute_distance(mic_positions[0], mic_positions[1], mode="torch")
+    cc, lag_indexes = gcc_phat(x1, x2, fs)
+    tdoa = lag_indexes[torch.argmax(torch.abs(cc))]
+    normalized_tdoa = normalize_tdoa(tdoa, mic_distances)
+
+    return normalized_tdoa
